Staging/requirements.txt
matplotlib
numpy
Pillow
fuzzywuzzy
python-Levenshtein